Move classes to different files and add PXPaymentPluginResult
