+++ MercadoPagoSDK/MercadoPagoSDK/Flows/PaymentFlow/PXPaymentFlowHandler.swift
@@ -1,0 +1,12 @@
+//
+//  PXPaymentFlowHandler.swift
+//  MercadoPagoSDK
+//
+//  Created by Eden Torres on 26/06/2018.
+//  Copyright © 2018 MercadoPago. All rights reserved.
+//
+
+import Foundation
+@objc public protocol PXPaymentFlowHandler: NSObjectProtocol {
+    @objc func showErrorScreen(message: String, errorDetails: String, retry: Bool)
+}
